Initial FT for 'My Lists'; adapt Persona integration
We work with pre-authenticated sessions in this FT. For that to
function, Persona mustn't try to log us out. Make the Persona
'onlogout' callback a NOP, and have Django logout happen only
through clicking the logout button.

NOTE: Still need to do a Persona logout when the button is
clicked, to avoid being logged back in again.